requirements.txt

transformers==4.31.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.2
torch==2.0.1
requests==2.31.0
pdfplumber==0.9.0
Pillow==10.0.0
